requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.5
langchain-ollama>=0.0.1
pydantic>=2.5.2
requests>=2.31.0
python-dotenv>=1.0.0
ollama>=0.1.5
fastapi>=0.104.1
uvicorn>=0.24.0
serpapi>=0.1.0 